+++ NEWS.md
@@ -4,6 +4,7 @@
 * The `show_logger_name` parameter has been added to the `mc_reshape_wide` function.
 * The `serias_name` label changed to `series` in `mc_plot_line`.
 * The progress bar is added to `mc_prep_crop`.
+* The `mc_prep_clean` step warning is fixed.
 
 # myClim 1.3.3
 * A performance issue in the `mc_prep_crop` function has been fixed.